Reland "Add SyncDecoupleAddressPaymentSettings to field trial config"
This is a reland of commit bfe9bb7ec091d71d6925526098f9d74fd0cb048a

The CL was reverted (crrev.com/c/5046251) because the test failed on
in the browser_tests_no_field_trial target. As the name suggest, in
this (apparently existing) target, the field trial config and hence
SyncDecoupleAddressPaymentSettings remained disabled. The test however
enabled SyncEnableContactInfoDataTypeInTransportMode, leading to a
feature combination that in practice is not possible (
SyncEnableContactInfoDataTypeInTransportMode is conditioned on and will
be launched prior to SyncDecoupleAddressPaymentSettings).

This caused the test to fail, because as of crrev.com/c/5023899:
- `syncing_or_addresses_in_transport_mode` is true
- `has_autofill_sync_type` is false, since AUTOFILL_PROFILE is
  not running (and not able to run) in transport mode.
Consequently, `syncing_or_addresses_in_transport_mode &&
!has_autofill_sync_type` became true, preventing credit card server
saves.

This CL fixes this problem by explicitly enabling both features in the
test in question.

Original change's description:
> Add SyncDecoupleAddressPaymentSettings to field trial config
>
> Bug: 1435431
> Change-Id: Ifed4ea2245e6ad7137bd88544509d130893d9354
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5024102
> Reviewed-by: Jan Keitel <jkeitel@google.com>
> Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
> Cr-Commit-Position: refs/heads/main@{#1226872}

Bug: 1435431
Change-Id: I48c8b8e5a54b2301e82a491a5f35bb04048573f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5049945
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1227864}
NOKEYCHECK=True
GitOrigin-RevId: 66e034d0eee181f46f0fa2b738f259dadfc77989