Exclude SettingsTest.A11yPage from the accessibility-linux builder
The linux-blink-web-tests-force-accessibility-rel bot is designed to run
content through the accessibility engine. Tests for many Settings pages
don't fall under this category, so are not part of this bot already.
SettingsTest.A11yPage is one of those, so this CL exclude this test
from that builder.

AX-Relnotes: n/a.
Bug: 1501815, 1442928
Change-Id: I04f60189f4e5bfb0f33365929a7c01e13e8dcd73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5037967
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1229549}
NOKEYCHECK=True
GitOrigin-RevId: 605ac284fe152b2166e60369cbac98574323ccdb