Enable PopulateVisitedLinkDatabase by default at ToT
We have received launch approvals for Stable 100% and Enabled-By-Default for this non-user-facing feature.

We are not launching on Android WebView right now, so we added an override to AwFieldTrials.

This CL also corrects an issue with the feature flag being enabled and disabled properly in relevant unit tests.

Launch Bug: launch/4259382

Bug: 1479889
Change-Id: I58af478366f8422a68aa3147344331c787e466d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5063449
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1230820}
NOKEYCHECK=True
GitOrigin-RevId: 022f4b9391a84f8d9a6dbf3b6dbdcaf8a5255daf