[infra] Remove support for and use of override_compile_targets field.
All uses of override_compile_targets are using the same value that would
be used for the test's compile targets already (the target in the test
field). Support for the field is being removed from the recipe. This
CL removes the field from the perf spec generator to reduce the
verbosity of the generated configs and removes the ability to set it
in the starlark code to avoid confusion.

Bug: 1506653
Change-Id: Id1d4dea2a670f13d63866940d8128a63f9e4014f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5077097
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1233063}
NOKEYCHECK=True
GitOrigin-RevId: 1b3f9cd1331bcf2e11d79bd6338af69e567fd29d