skylab builder: Increase the number of shard for chrome_all_tast_tests
This CL increases the number of shards as follows:
- jacuzzi&trogdor: 10 -> 20
- other: 10 -> 15

We decreased the number of shards when we launched the skylab builders.
As we have increased the number of tests, the execution time has been
longer, so that this CL increases the number of shards.

Especially jacuzzi and trogdor are slow and sometimes hitting the timeout, so that they are increased to 20.

BUG=b/314884667
TEST=None

Change-Id: I28d7a3abe2a2a58d353007ffd88f3ea6bbf9d864
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5086972
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Queue: Xinan Lin <linxinan@chromium.org>
Reviewed-by: Xinan Lin <linxinan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1233818}
NOKEYCHECK=True
GitOrigin-RevId: cfc8ff9c6fa911a3ca7ecf65d8e98951da2466b5