Add an experiment of framerate throttling for single-video cases
Currently for single-video pages that have only the video is updating,
the framerate toggling is not enabled. This CL add an experiment to
enable it behind a feature for perfect cadence.

This also removes the obsolete checks of frame sinks with no
preference interval in FrameRateDecider.

Bug: 1478671
Change-Id: If40ba460964c745c95caf4474255f9f9b3015a40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4837864
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Jiahe Zhang <jiahe.zhang@intel.com>
Cr-Commit-Position: refs/heads/main@{#1236139}
NOKEYCHECK=True
GitOrigin-RevId: f7ad8a03b0a299786604415a1567ddedbc97e971