Reland [Privacy Sandbox] Turn attestation default-allow on.
Original CL crrev.com/c/5116010 was reverted in crrev.com/c/5120393.

The root cause is the test assumes the behavior is default deny, but
we have changed the behavior to default allow, temporarily.
There is also a problem in the test setup where we should delete the
component installation folder both before and after the test.

Note this CL only fixed the failing test. We would like this to submit
as soon as possible because the finch CL that turns the default-allow
on has already been merged.

Create crbug.com/1511477 to track the task of disabling default-allow
feature for all tests that use attestation. This is because the
default-allow is only a temporary fix.

Bug: 1500636, 1448875
Change-Id: Ie72aeb117325a532dd06b4f73fd87fdc0b767013
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5117203
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#1237587}
NOKEYCHECK=True
GitOrigin-RevId: 455c66b7710ee9cd176df2cdbed6a465c8a6c494