Remove Flag and Entry points to GVR
The WebXR Cardboard feature has been shipped for two milestones now with
no regressions. This change removes the now-expired UI flag that would
allow disabling it, and also begins the work of removing the backend
that it replaced (GVR), by removing the appropriate flag entry points
for it, as well as the ability to create it as a provider. This also
deletes all tests for the GVR code and paves the way for removing the
GVR backend which should no longer be used.

Bug: 1510066
Change-Id: Iefe7021b84aa52eaaf5096529e93433188c5b58c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5105042
Reviewed-by: Brandon Jones <bajones@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237770}
NOKEYCHECK=True
GitOrigin-RevId: 83d415aec1f093dfd79b04e516666e4691c7cec2