[Lacros a11y] Sticky keys browsertests work in Lacros
Gets most sticky keys browsertests that require the browser
to work in Lacros using the new Autoamtion testing API,
except one that is left as a TODO. Filters a few others
that do not require the browser from running in Lacros as
that would waste resources.

Installs Select to Speak to have a background extension that
can access Automation; Select to Speak is not triggered
during the test. This matches the focus highlight tests.

Bug: b/301475127
Change-Id: I21085e2c62ea8b26da4fe745e9064c34bfc4c35b
Ax-Relnotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5124051
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237776}
NOKEYCHECK=True
GitOrigin-RevId: 2240ae79ce5605c1ed5f41cc657f5b082aaac39f