[accessibility] Introduce ScopedAccessibilityMode
A ScopedAccessibilityMode instance can be created (via
content::BrowserAccessibilityState::CreateScopedModeFor...) to apply a
set of AXMode flags to all WebContentses in the browser process, to all
WebContentses associated with a specific BrowserContext, or to a
specific WebContents. When created and destroyed, the union of the mode
flags in all live instances are applied to all relevant WebContentses.
This mechanism allows different controllers of accessibility modes to
enable/disable mode flags without interfering with one another. In this
stage, the only controller using the new mechanism is
BrowserAccessibilityStateImpl.

Today's callers of methods like {Add,Remove}AccessibilityModeFlags and
NotifyAddAXModeFlags will be migrated to using ScopedAccessibilityMode
in subsequent CLs.

Bug: 1470199
Change-Id: I22a2aac3e70743b65afcafcd471d2748cc0281c0
AX-Relnotes: n/a.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5018479
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244200}
NOKEYCHECK=True
GitOrigin-RevId: a8d2f11254422d03c66cd7a0a0688722e9784fe1