[Chrome Labs] Fix dialog corner rounding and title spacing
This CL moves the top insets from the scroll view contents
to the scroll view itself. This prevents the menu from scrolling too
close to the top label which fixes the issues described in the bug.

This CL also adds separate logic for non-Mac and Mac platforms. Mac will have an additional border at the bottom of the scroll view since
the scroll bar is painted on top of the bubble on Mac which causes
gaps with the rounded corners. This looks a bit strange on Windows so
we will keep the current behavior for non-Mac platforms.

Bug: 1489424
Change-Id: Idd27436a05f89d21c34b4622085201bc44c15291
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4980003
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Elaine Chien <elainechien@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244437}
NOKEYCHECK=True
GitOrigin-RevId: dfa97398a3f71f15106dce662859912404d3a8d8