Move vaapi_unittest configuration to a mixin args
We want to run the unittest with a different envionment variable
configuration. For this purpose the gtest args needs to be set by test
suite or mixin instead of the tests.gtest_test, allowing to change it at
the test suite level.

Bug: b:311608250, b:315718813
Change-Id: I2e023fa23439350e183252a42676eeec3fbdcdc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5143749
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244452}
NOKEYCHECK=True
GitOrigin-RevId: f079d730f5422aacb1a2330b94d559109e8dfbe1