Shard Cronet instrumentation tests
The P50 of this suite is from 650 ~ 850 seconds which maps to (10 to 14 minutes), after taking a close look at one of the runs (https://ci.chromium.org/ui/p/chromium/builders/ci/android-cronet-x86-dbg-10-tests/53554/overview), the shard runtime was 16 minute while the overhead was just 11 second. The tests are slow because they are executed in sequential order, sharding will result in a significant speed-up.

I've tried to create multiple (3) emulators and test this out locally and it seems to have caused a speed-up, unfortunately. There is no experimentation framework for the eng-prod so this will be a best-basis effort to determine if this change has caused improvement or not through a dashboard.

The number of shards chosen is an ad-hoc choice more than a scientific one, The average runtime is 12 minutes, so 3 shards should bring it down to (4-6~) minutes. Increasing the number of shards results in minimalistic change.

Change-Id: I22d569e7f7b4215910c211daf8ab659c8c1bc759
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5177513
Reviewed-by: Haiyang Pan <hypan@google.com>
Commit-Queue: Haiyang Pan <hypan@google.com>
Auto-Submit: Mohannad Farrag <aymanm@google.com>
Cr-Commit-Position: refs/heads/main@{#1245395}
NOKEYCHECK=True
GitOrigin-RevId: 6e446ee2be5fcd226598010b4ae64c789d6a4b37