[ios] Fix the broken voiceover in debugging envir.
The accessibility is not working for the web view content in debugging
mode when the Screentime feature is on, likely an Apple issue. Fix by
not compiling the code for the feature.

The code is not removed yet because I want to see whether this change
reduces the Screentime-involved crashes to 0, to be more sure about that
it is our integration that is introducing the crash, versus something
else.

Fixed: 1508811

Change-Id: I346dc0f72b7080e81fbda089ee9418757af7e25d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5186575
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1245806}
NOKEYCHECK=True
GitOrigin-RevId: 5f4521bd03b86d9a6a178c330e2eb3b01e619427