Add testing config for DeprecateUnloadAllowlist.
This list is the top 50 sites according to Wikipedia.

In 40 cases, the naked domain will redicted to www. and so www. is used.
In 6 cases, www. redirects to the naken domain, so the naked domain is used.
In 4 cases neither redirects, so both are used.

Bug: 1518694
Change-Id: I088d1c97bdfd55611ab79d6c197f24af5ba1e408
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5223872
Reviewed-by: Ming-Ying Chung <mych@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1251245}
NOKEYCHECK=True
GitOrigin-RevId: 5ed4450f68d99bece3824dd5d964acafe96d82d8