cleanup: Remove remnants of BrowserPeriodicYieldingToNative
The crrev.com/1111568 removes BrowserPeriodicYieldingToNative with a
message: "Since there are no immediate plans to ship this feature".

Remove a few members of the BrowserUIThreadScheduler. They were only
used under BrowserPeriodicYieldingToNative. Not used any more, not even
in tests.

Leaving the |scroll_state_| in place because it sinks in traces, where
it might be useful.

Bug: 1215076
Change-Id: Ibbd38c3b082436e24614a9c10341fe865984ea3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5238435
Reviewed-by: Omar Elmekkawy <mekk@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253377}
NOKEYCHECK=True
GitOrigin-RevId: 409133287292c75a1ee384b8cf2fa6ddaa97f7df