Stop tests on chromeos-betty-arc-r-chrome
We have ran the tests of betty-arc-c on DUTs in Chrome browser labs,
because GCE VM doesn't support running ARCVM.

But we are removing the DUTs from the lab so that the test are unable
to run. Until the VM supports ARCVM, we disables the tests.

BUG=b:323287977
TEST=None

Change-Id: Ia975ecf16a5a743646ff18ea69d7e811e3b5d95a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5256563
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Xinan Lin <linxinan@chromium.org>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254941}
NOKEYCHECK=True
GitOrigin-RevId: 662360e17b494bde0c7bdb31d351f85b13c98fa1