Create experimental JNI implementation for Android accessibility
Start an experiment that takes a different approach to the JNI
calls for the Android accessibility code. Instead of passing a
large object over the boundary, pass an id and use the existing
cache in the Java-side code to reference the object. This
experiment may not be pushed beyond Beta depending on how the
initial data looks.

AX-Relnotes: N/A
Bug: N/A
Change-Id: I44ceb6e7ade9c03fd221b8d3369d51bbf2ee8744
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253290
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1255111}
NOKEYCHECK=True
GitOrigin-RevId: a8ebd0515a01a1d5e8e3fb7079dc8064334dc043