[omnibox] Reports # of zero-prefix suggestions in Autocomplete session.
Omnibox currently reports the number of zero-prefix suggestions shown in
searchboxstats (gs_lcrp=) only if the user selects a zero-prefix
suggestion. In other words, omnibox does not take into account the
zero-prefix suggestions previously shown to the user if they end up
typing and selecting a prefixed suggestion. This is incorrect and makes
a number of otherwise insightful Bsb metrics useless. See
go/chrome-zps-metrics-101 for details on the affected metrics.

This CL introduces the concept of a session in the AutocompleteResult
which is reset when the AutocompleteController resets the
AutocompleteResult - i.e., on popup closing. This is not to be confused
with when the AutocompleteController clears the matches - e.g. on the
last async pass. This allows the session metadata to be carried over
between the "old results" (results from the previous query) to the
"new results" (results from the current query) until the popup is closed. The session metadata is at the moment restricted to the number of zero-prefix suggestions. If it proves to work well, it can also
include the "previous suggestions" in the future.

Pros:
The scope of changes is restricted to the Autocomplete classes allowing
it to work across all platforms (verified by Mobile teams).

Alternative considered:
Building on top of the existing concept of
AutocompleteController::ResetSession() and
OmniboxTriggeredFeatureService. Unfortunately that definition session
does not work for zero-prefix suggestions as it defines a new session
as when the user starts editing the input. Changing that definition
involves making changes to the Omnibox UI classes (not used by Chrome
Android) and to the Mobile platforms codes which takes a much larger
amount of investment.

Future work:
Exploring whether this approach can replace
OmniboxTriggeredFeatureService and whether it can be consolidated with
AutocompleteControllerMetrics for a single more manageable and better
tested metrics. Eventually bringing the concept of session metrics to
OmniboxLog can enable a variety of potentially useful session metrics
to omnibox histograms.

Bug: 1307142
Change-Id: I989221bd8bb2cd939d5c85227726a156832c5d9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5280536
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1258592}
NOKEYCHECK=True
GitOrigin-RevId: 90235ef6cbd176b09101d2d3d0006684eca1ce83