Reland "Reland "[fieldtrial_testing_config] Enable IdentifiabilityStudyMetaExperiment""
This is a reland of commit 13621267e1e35d70b52a5f00b078b24932f1503a

The browser test was failing on the bot because the behaviour was
depending on the fieldtrial_testing_config, which is enabled only for
non-chrome-branded bots (which I did not know). A comment in the test
was suggesting that the test was indeed checking whether the
fieldtrial_testing_config was being picked up correctly, which is why
I was assuming the test would work. But that was not true, the test
was working correctly just because the features and features param
where initialized correctly in the test class constructor with a
ScopedFeatureList. Now I also locked in the state of the new feature
in the test class constructor within the ScopedFeatureList, and this
should work correctly on all bots.

Original change's description:
> Reland "[fieldtrial_testing_config] Enable IdentifiabilityStudyMetaExperiment"
>
> This is a reland of commit 416136ed0d74af8db680d05d2ceda1ac669bd93c
>
> Original change's description:
> > [fieldtrial_testing_config] Enable IdentifiabilityStudyMetaExperiment
> >
> > This CL adds a new entry to the fieldtrial_testing_config for enabling
> > the IdentifiablityStudyMetaExperiment feature.
> >
> > Bug: 324271182
> > Change-Id: I3f676943b5ac181f83b3e5b6b210648728649175
> > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5315258
> > Reviewed-by: Mike West <mkwst@chromium.org>
> > Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
> > Cr-Commit-Position: refs/heads/main@{#1263923}
>
> Bug: 324271182
> Change-Id: I9aefaad7e4376d7e94618be590123d992c4b8a1d
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318252
> Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
> Reviewed-by: Mike West <mkwst@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1266360}

Bug: 324271182
Change-Id: Ib254246af8890c28029cfc656c91a2f98baa4923
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5332474
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1268531}
NOKEYCHECK=True
GitOrigin-RevId: 1597df286d180e46e0c2cd9862abb149553ab685