Remove stale configuration of `chromeos-eve-arc-r-chrome`
The builder named `chromeos-eve-arc-r-chrome` had been removed, but
there was the stale testing configuration of the builder. I assume
we had forgotten to remove it. This CL clean it up.

Also, by removing the testing configuration, `chromeos_device_gtests`
is no longer referenced by any builder, so that this CL removes it as
well.

This is a clean-up CL, so that is should not change any behavior.

Bug: None
Test: None

Change-Id: I007a041864df23ffd661bcf2f7b6aff885c5d26a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5343540
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Xinan Lin <linxinan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1268815}
NOKEYCHECK=True
GitOrigin-RevId: 55ba7817dac4edeb887bcbb35c63d525796b0f98