[iOS] Enable feed containment fieldtrial and fix test
Adds feed containment to fieldtrial testing config and fixes test that
was failing. The failure was caused by the margins being added to the
collection view, which affects the width when comparing to the pinned
omnibox. Instead, we now check the width of the NTP parent view which is
not affected by the margins.

Bug: 40285257
Change-Id: I560c7a4d447440560db86be080800c5e6c523044
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5348384
Commit-Queue: Adam Arcaro <adamta@google.com>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1269145}
NOKEYCHECK=True
GitOrigin-RevId: 9ef2e4a5f0dabd04d58bfa2685067f2ffae2b0bc