Add ReadAndDiscard option to URLLoader
Add a new URLLoader option which requests that it read and discard the
response body itself rather than writing it to a data pipe. This is more
efficient when the consumer doesn't need the body. The URLLoaderClient
will be passed an empty data pipe handle in OnReceiveResponse().

The new option is only implemented for network::URLLoader. Because there
are over 500 implementations of URLLoaderFactory::CreateLoaderAndStart()
it is not possible to implement the option in all of them. Instead,
implementations of URLLoaderClient must be prepared to receive and
discard a body even if they requested the option.

The option is currently disabled when ORB is enabled as we don't need it
in that case and it complicates the implementation.

Add a feature flag "SimpleURLLoaderUseReadAndDiscardBodyOption" which
causes SimpleURLLoader to set the kURLLoadOptionReadAndDiscardBody
option when creating a URLLoader in the DownloadHeadersOnly() method. If
the URLLoader supplies no body in this case the body is considered to be
successfully downloaded.

Modify SimpleURLLoader's unit tests to test all three cases of flag
enabled, flag disabled, and flag enabled but the option not implemented
by the URLLoader.

Also add a fieldtrial testing config for the feature flag so that it
will be enabled in browser tests to give additional coverage.

Change-Id: I1d609f679334f8650dac27a51bfc90ec972c784c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5348681
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1271270}
NOKEYCHECK=True
GitOrigin-RevId: dee3d6e2e358f365300ff850af2ede1079b7fcd5