Add a testsuite for preuprev tests
This CL adds a test binary to execute the tast tests of group:cq-medium, that is a subset of the essential tast tests (see http://go/cq-medium).

Also this CL adds a testsuite to execute the essential tests including
cq-medium, that will be called from the chromeos preuprev builders.

With this CL, the preuprev builders run the following tests:
- The builders except the linux builder
  - chromeos_cq_medium_tast_tests
  - chromeos_integration_tests
  - chromeos_device_only_gtests (base_unittests)
- The Linux builder
  - linux_chromeos_gtests

Files manually changed:
- chromeos/BUILD.gn
- infra/config/targets/basic_suites.star
- infra/config/targets/binaries.star
- infra/config/targets/compound_suites.star
- infra/config/targets/matrix_compound_suites.star
- infra/config/targets/tests.star
- testing/buildbot/waterfalls.pyl
(the other changes are generated by the script)

Bug: b/286503577
Test: None
Change-Id: If338248d06387e5ce2856a411a8445463c6e1564
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5372787
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1274061}
NOKEYCHECK=True
GitOrigin-RevId: ee7d8a239ffd3d4d92909ed61c94aa1103fd0176