[PWA] Enable WebAppUniversalInstall via field trial testing configs
This CL enables the WebAppUniversalInstall feature in the field
trial testing configs, and does some minor fixes:

1. For non-installable (DIY) apps, the `Open in PWA` menu item
that should show up in place of `Install as app` item does
not show up until a few re-renders of the app model menu. To
fix that, the app_menu_model tracks whether the current
site url already has an existing installed app, and prevents
the install button from showing up. The hypothesis is that
this happens because the AppBannerManager pipeline is slow
enough that it cannot finish before the first few renders,
something that was discussed for testing in [1]

2. The simple install dialog view has a different name once
the flag is switched on, so WidgetObservers had to be updated
to use that name.

2. The INSTALL_IDC_PWA command based menu item entry is always
available in the three dot menu now (even for non installable
sites), so tests previously verifying that the install menu option
wouldn't show up is not fixed.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5320248/comment/b6bb665b_fca91ec2/

Bug: 326611791, b/329431634
Change-Id: I751e77a0a5ebe8dae6024a725630489119372f94
Include-Ci-Only-Tests: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5369151
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1274571}
NOKEYCHECK=True
GitOrigin-RevId: 48cce2cf43e1957d0f7966ba4cbab601b6113b68