Reland "[Permission] Creates a pixel test for camera preview in permission prompt."
This is a reland of commit 0eb295185782753cf75a804e3ef5ab2e908bae81

The original submit failed because there is a image difference between the golden image and the test image, which require to be triaged.
The difference was some pixel differences on the border (like the shadow). Overall, everything looks fine.
The result is triaged and a reland is created without changing anything in the code.

Original change's description:
> [Permission] Creates a pixel test for camera preview in permission prompt.
>
> Bug: 296947420
> Change-Id: I6da1602668c9ed711a298d0116a1b4880de982e8
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5362800
> Reviewed-by: Ahmed Moussa <ahmedmoussa@google.com>
> Reviewed-by: Bryant Chandler <bryantchandler@chromium.org>
> Commit-Queue: Clifford Cheng <cliffordcheng@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1275448}

Bug: 296947420
Change-Id: Iace3f2256b778563d6a496677ea83b3edae026c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5383156
Commit-Queue: Clifford Cheng <cliffordcheng@chromium.org>
Reviewed-by: Bryant Chandler <bryantchandler@chromium.org>
Reviewed-by: Ahmed Moussa <ahmedmoussa@google.com>
Cr-Commit-Position: refs/heads/main@{#1275809}
NOKEYCHECK=True
GitOrigin-RevId: 9dfa6101d1a0aacd763e8a2f5b4d508a08543d8d